Support script-to-script includes
