+++ crates/clitest-lib/src/parser/v0/mod.rs
@@ -1,0 +1,4 @@
+mod parse;
+mod segment;
+
+pub use parse::parse_script;
