+++ tests/v0/basic_parse.cli
@@ -50,4 +50,4 @@ c
 !!!
 
 $ printf "%%{}\n"
-! %{}
+? %\{\}
